requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
mysql-connector-python==8.3.0
selenium==4.17.2
Flask-Cors == 4.0.0